Move resizeImage function to new utils file
