+++ src/main/webapp/app/quiz-statistics/drag-and-drop-question-statistic/drag-and-drop-question-statistic.component.ts
@@ -8,11 +8,11 @@ import { DragAndDropQuestionUtil } from '../../components/util/drag-and-drop-que
 import { ArtemisMarkdown } from '../../components/util/markdown.service';
 import { DragAndDropQuestion } from '../../entities/drag-and-drop-question';
 import { DragAndDropQuestionStatistic } from '../../entities/drag-and-drop-question-statistic';
-import { QuizQuestionType } from '../../entities/quiz-question';
 import { DropLocation } from '../../entities/drop-location';
 import { ChartOptions } from 'chart.js';
 import { createOptions, DataSet, DataSetProvider } from '../quiz-statistic/quiz-statistic.component';
 import { Subscription } from 'rxjs/Subscription';
+import { resizeImage } from 'app/utils/drag-and-drop.utils';
 
 interface BackgroundColorConfig {
     backgroundColor: string;
@@ -106,7 +106,7 @@ export class DragAndDropQuestionStatisticComponent implements OnInit, OnDestroy,
     }
 
     @HostListener('window:resize') onResize() {
-        this.resizeImage();
+        resizeImage();
     }
 
     getDataSets() {
@@ -360,20 +360,5 @@ export class DragAndDropQuestionStatisticComponent implements OnInit, OnDestroy,
         } else {
             return null;
         }
-    }
-
-    /**
-     *  The background image for drag and drop quizzes is dynamically adjusted through scss, therefore we have to resize the click layer to the same width and height as the background image
-     */
-    resizeImage() {
-        /* set timeout as workaround to render all necessary elements */
-        setTimeout(() => {
-            const image = document.querySelector('.drag-and-drop-quizStatistic-picture img') as HTMLImageElement;
-            const clickLayer = document.getElementsByClassName('click-layer').item(0) as HTMLElement;
-            if (clickLayer) {
-                clickLayer.style.width = image.width + 'px';
-                clickLayer.style.height = image.height + 'px';
-            }
-        }, 100);
     }
 }
