+++ src/main/webapp/app/quiz/participate/drag-and-drop-question/drag-and-drop-question.component.ts
@@ -7,6 +7,7 @@ import { DropLocation } from '../../../entities/drop-location';
 import { polyfill } from 'mobile-drag-drop';
 import { scrollBehaviourDragImageTranslateOverride } from 'mobile-drag-drop/scroll-behaviour';
 import { SecuredImageComponent } from 'app/components/util/secured-image.component';
+import { resizeImage } from 'app/utils/drag-and-drop.utils';
 
 // options are optional ;)
 polyfill({
@@ -80,7 +81,7 @@ export class DragAndDropQuestionComponent implements OnChanges {
     constructor(private artemisMarkdown: ArtemisMarkdown, private dragAndDropQuestionUtil: DragAndDropQuestionUtil) {}
 
     @HostListener('window:resize') onResize() {
-        this.resizeImage();
+        resizeImage();
     }
 
     ngOnChanges(changes: SimpleChanges): void {
@@ -116,7 +117,7 @@ export class DragAndDropQuestionComponent implements OnChanges {
     changeLoading(value: string) {
         this.loadingState = value;
         if (this.loadingState === 'success') {
-            this.resizeImage();
+            resizeImage();
         }
     }
 
@@ -293,20 +294,5 @@ export class DragAndDropQuestionComponent implements OnChanges {
      */
     countCorrectMappings(): void {
         this.correctAnswer = this.question.dropLocations.filter(dropLocation => this.isLocationCorrect(dropLocation)).length;
-    }
-
-    /**
-     *  The background image for drag and drop quizzes is dynamically adjusted through scss, therefore we have to resize the click layer to the same width and height as the background image
-     */
-    resizeImage() {
-        /* set timeout as workaround to render all necessary elements */
-        setTimeout(() => {
-            const image = document.querySelector('.background-area jhi-secured-image img') as HTMLImageElement;
-            const clickLayer = document.getElementsByClassName('click-layer').item(0) as HTMLElement;
-            if (clickLayer) {
-                clickLayer.style.width = image.width + 'px';
-                clickLayer.style.height = image.height + 'px';
-            }
-        }, 100);
     }
 }
